AP; InstallWinGet: hash packages before installing also remote update to Windows Package Manager v0.3.11201 Preview

+++ CW-GUI.ps1
@@ -499,18 +499,26 @@ $InstallWinGet.Add_Click( {
     Import-Module BitsTransfer
 	$result = Test-NetConnection github.com
 	if( $result.PingSucceeded ) {
-        Write-Host "Downloading required files..."
+        Write-Host "Downloading WinGet installation packages..."
         Start-BitsTransfer https://github.com/CleanWin/Files/raw/main/Microsoft.DesktopAppInstaller_8wekyb3d8bbwe.appxbundle
         Start-BitsTransfer https://github.com/CleanWin/Files/raw/main/Microsoft.VCLibs.140.00.UWPDesktop_14.0.29231.0_x64__8wekyb3d8bbwe.Appx
-        Write-Host "Installing Windows Package Manager (WinGet)..."
-        Add-AppxPackage -Path .\Microsoft.DesktopAppInstaller_8wekyb3d8bbwe.appxbundle -DependencyPath .\Microsoft.VCLibs.140.00.UWPDesktop_14.0.29231.0_x64__8wekyb3d8bbwe.Appx
-        Remove-Item Microsoft.DesktopAppInstaller_8wekyb3d8bbwe.appxbundle
-        Remove-Item Microsoft.VCLibs.140.00.UWPDesktop_14.0.29231.0_x64__8wekyb3d8bbwe.Appx
-        Write-Host "Installed Windows Package Manager (WinGet)."
-	} 
-	else {
-	  Write-Host "Could not connect to the internet. WinGet won't be installed."
-	}
+		# Hash the files, if hashes match, begin installation or write warning.
+        $filehash1 =(Get-FileHash "Microsoft.DesktopAppInstaller_8wekyb3d8bbwe.appxbundle" -Algorithm SHA256).Hash
+        $filehash2 = (Get-FileHash "Microsoft.VCLibs.140.00.UWPDesktop_14.0.29231.0_x64__8wekyb3d8bbwe.Appx" -Algorithm SHA256).Hash
+        if ( ($filehash1 -eq "859999DFD7FAAE1FB9AC56760736301950C937EA5BA4F66387CEEE7A76CB3D00") -and ($filehash2 -eq "6602159C341BAFEA747D0EDF15669AC72DF8817299FBFAA90469909E06794256") ) {
+            Write-Host "Successfully verified package hashes."
+            Write-Host "Installing WinGet..."
+            Add-AppxPackage -Path .\Microsoft.DesktopAppInstaller_8wekyb3d8bbwe.appxbundle -DependencyPath .\Microsoft.VCLibs.140.00.UWPDesktop_14.0.29231.0_x64__8wekyb3d8bbwe.Appx
+            Remove-Item Microsoft.DesktopAppInstaller_8wekyb3d8bbwe.appxbundle
+            Remove-Item Microsoft.VCLibs.140.00.UWPDesktop_14.0.29231.0_x64__8wekyb3d8bbwe.Appx
+            Write-Host "Installed Windows Package Manager (WinGet)."
+        }
+        else {
+            write-host "Could not verify package hashes. WinGet won't be installed."
+        }
+    else {
+        Write-Host "Could not connect to the internet. WinGet won't be installed."
+    }
 })
 
 $Winstall.Add_Click( {
